pyproject.toml
[project]
name = "refinery"
version = "0.1.0"
description = "The Document Intelligence Refinery - Multi-Stage Extraction Pipeline"
authors = [
    { name = "FDE Candidate", email = "candidate@example.com" }
]
requires-python = ">=3.10"
dependencies = [
    "pydantic==2.7.1",
    "pdfplumber==0.11.4",
    "PyYAML==6.0.1",
    "pytest==8.2.0",
    "requests==2.31.0"
]

[project.optional-dependencies]
vector = [
    "chromadb==0.5.5"
]
agent = [
    "langgraph==0.2.59"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
